--- a/scripts/train_two_player_emb.sh
+++ b/scripts/train_two_player_emb.sh
@@ -16,4 +16,4 @@ python train.py --name lsgan_emb \
        --D_iters 1 \
        --score_name muse-csls-en \
        --print_freq 2000 --display_freq 2000 --score_freq 10000 \
-       --save_latest_freq 100000 --save_giters_freq 100000
+       --save_latest_freq 100000 --save_giters_freq 100000 --most_frequent 75000