- 只取前most_frequent(MUSE取75000)個字train，accuracy很快收斂到很好的解
- MUSE的epoch size是用幾個word來算，和我們用iteration的算法不一樣(參考[unsupervised.py](https://github.com/facebookresearch/MUSE/blob/master/unsupervised.py) )
- 所以以每秒處理幾個word來看，我們不會比MUSE慢(9000samples/s v.s. 7000samples/s 9000samples/s v.s. 7000samples/s 在人工智慧學校上)